% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formats.R
\name{pdf_grid3}
\alias{pdf_grid3}
\title{PDF Output for GRID3 Reports}
\usage{
pdf_grid3(
  toc = FALSE,
  toc_depth = 2,
  number_sections = FALSE,
  latex_engine = "pdflatex",
  extra_tex = "",
  ...
)
}
\arguments{
\item{toc}{TRUE/FALSE (default TRUE). Display a table of contents}

\item{toc_depth}{the depth of the table of contents. Default is 2}

\item{...}{additional arguments which can be parsed by the word_document function}
}
\description{
A custom report style to fromat the output of R Markdown reports.
 Uses bookdown's \code{pdf_document2} as the base format to allow
 cross-referencing and other improvements.
}
\details{
Lower level LaTeX YAML settings are specified in an external file. This
 controls the document paper size, margins, and other layout options.
}
\seealso{
\code{\link[rmarkdown]{pdf_document}}
}
\author{
Michael Harper
}
