% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formats.R
\name{word_grid3}
\alias{word_grid3}
\title{Word Output for GRID3 Reports}
\usage{
word_grid3(toc = TRUE, toc_depth = 2, ...)
}
\arguments{
\item{toc}{TRUE/FALSE (default TRUE). Display a table of contents}

\item{toc_depth}{the depth of the table of contents. Default is 2}

\item{...}{additional arguments which can be parsed by the word_document function}
}
\description{
Uses bookdown package to allow cross-referencing. Uses the styles reference
 file `mystyles.docx` stored in directory `inst/rmarkdown/templates/elements`
 to define custom styling.
}
\seealso{
\code{\link[rmarkdown]{word_document}}
}
\author{
Michael Harper
}
